Enable easy configuration of server API location
Moves API location to a location separate from NGCHM viewer, so
that it simplifies writing rules for front-end proxies etc.

Define that location in this small static file so that it can be
changed easily.  (e.g. by having proxy redirect to a customized file.)

Also include location of linkout customization script here as well.