Only render summary heatmap if data or parameters have changed
Creates an object containing all the parameters that affect rendering
of the summary heat map.  Using JSON.stringify to turn it into a string
can comparing it to the previous value is supposed to be faster than
a complicated object traversal in javascript.  Even if it's only
competitive, the simplicity is worth it.

The other significant change is for the NEWDATA message to include all the
details about the updated tile.