Send NEWDATA messages to all alternate levels
For example, in a small map there is only a summary level.  Summary level
is therefore used as the alternate for detail level.  Detail level is also
used as the alternate for horizontal ribbon and vertical ribbon level.  So
this commit ensures that a task listening for changes to tiles at the
detail level will be notified when the summary tile is received.