Speed up calculation of many label widths by reducing forced reflows
All temporary labels that need a new size calculated are added to the
document before forcing a reflow.  Instead of a reflow per label, we get
one per label group (row labels, column labels, covariate labels, etc.)

The effect is specific to situations when a large number of labels need to
be calculated, including startup (to a modest degree) and when panning
to new areas of the map when zoomed out far enough for a lot of labels
to be displayed.