Implement RESET button & fix bugs
This commit implements the RESET button on the Gear Dialog, fixes
some bugs, and makes code and UI improvements on the Gear Dialog.
Most of the changes are in Linkouts.js, with two exceptions in
Dendogram.js and NGCHM_Util.js noted below, and some changes to
the css in NGCHM.css.

Bug fixes:

- Selection of Scatter Plot points by clicking dendograms
  - this is the change in Dendogram.js
- Disallows selection of non-numeric covariates for Coordinates in
  Gear Dialog
- Selection of points between two Scatter Plots

Gear Dialog UI improvements

- In choosing groups from discrete covariates, limit the height of
  the selection box and allow for scrolling (sometimes the list of
  discrete covariates is very long).
  - part of this is the change in NGCHM_Util.js. This change of
    the wheel event listener was made to allow for the wheel
    action to scroll in the Gear Dialong and NOT cause zooming of
    the heat maps.
- The RESET button works now

Gear Dialog code improvements

- Refactor of existiong code from my Pull Request #33 to better fit
  into the existing code structure