Implementation for different axis in Gear Dialog
The implementation for switching the axis in the gear dialog
was not fully implemented. This is the first drop down
in the dialog, called 'Point Axis' in the Scatter Plot,
and 'Group Axis' in Pathway Mapper.

This commit adds that implementation; however further testing
is required.

There is also the start of implementation of saving the last
data passed to the plugin (so that the RESET button will work
for both axes). This is still in progress.