Give system message for pane manipulation
Instead of the simple 'confirm' box, use a system message to
collect user input on the decision to lose PathwayMapper state
when dividing panes.

Also, when closing a pane, if that closure will result in
PathwayMapper being reset, give a warning and opportunity to cancel.

(this functionality can be refactored for improvement: bascially
the same promisePrompt function is used in two places.
TODO: move that functionality to UserHelpManager)