Implement framework for dropping Linkouts onto NG-CHM window
User can drag and drop an item onto a 'droptarget' near the hamburger
menu.  Normally the drop target is invisble, but it turns pink and
displays 'DROP HERE' when the mouse is over it.

The only item recognized is a JSON-formatted link out specification,
but dropped it just outputs a console message in this commit.