Generalize warning about PathwayMapper state loss
In commit 63c6c837538fefa2e8cfcf20b49f982edb58df1c we added code to warn
the user if they were going to lose the state of PathwayMapper (a state
they may have spent a long time creating), and give them a chance to
cancel

This same issue occurs for all the plugins. For the ScatterPlot plugins,
it's less of a major concern because the information lost is typically
less time consuming to regenerate. However it's still a concern, and
noted in issue:
https://github.com/MD-Anderson-Bioinformatics/ScatterPlotPlugin3D/issues/32

This commit makes the warning message more general, but keeps the
language for the original PathwayMapper for that particular plugin.