Allow for NO GRAB/SHOW option in gear menu
Some plugins, for example the PCA plus plot plugin, do not want to allow
the user to use GRAB/SHOW for selecting data to plot. Updated the
addCovariateOptions function such that if the 'selectedElementsOption'
text input is null, then the GRAB/SHOW option is not added to the
dropdown.

Also added some comments to try and help clarify this part of the code.