Only show GRAB/SHOW if value starts with 'Selected '
The original code assumes all plugins want to allow the user to grab
rows or columns to use as data for the plugin, and therefore set things
up so that the last entry in the coordinate/covariate dropdowns was to
display the GRAB/SHOW buttons.

However some plugins, like the PCA Plus Plot, do not want to give the
user this option. Made this change to allow for these plugins that do
not want GRAB/SHOW buttons.

This adds another brittle element to the code: if we change the text
value of 'Selected rows/columns for __', the GRAB/SHOW buttons will
never be displayed.