Be less specific determining plugin StateMan
I have a recurring issue with naming of plugins regarding naming.
Example: Plugin name in custom.js is '2D ScatterPlot'. But when data
is sent to the plugin via the gear menu, the pane header gets the text
'2D ScatterPlot. Heat Map Columns'. So when saving the panel
layout, the title saved is '2D ScatterPlot. Heat Map Columns'.

For the moment, making the more flexible and potentially error-prone
choice in this commit. However this needs improvement.

If someone decides to call the 2D scatter plot
'Two-dimensional scatter plot' in their custom.js, this whole thing
will break.