Add function to save posted plugin data mapConfig
Added a function to save the data sent to a plugin (e.g. when the user
clicks 'APPLY' in the gear dialog) to the mapConfig.json. This data
will then be saved in mapConfig.json when the user saves the .ngchm
file. Call this function after posting data to plugin.