Make promises for creating access window
Added a set of parallel functions in MatrixManager for getAccessWindow
and its sub functions. These new functions return Promises that resolve
when all tiles for the access window are available in the tile cache.

This was done because we sometimes have the following issue: when a
function such as getAccessWindow().getValue() is called before the
needed tile is available in cache, a summary value is returned. This
summary value is not always desired: for example if the returned value
is used for calculating mean, t-test, etc., a summary value yields
incorrect results.