Use promises to get statistical test data
In the code that gets the summary statistics for mean, std dev, t-test,
use the Promise version of getAccessWindow to ensure values returned
from getValue() are NOT summary values.

In order to accomplish this, a few other functions were changed to
return promises.

All of this this was done to address issue #17 in the PathwayMapper
project:
https://github.com/MD-Anderson-Bioinformatics/pathway-mapper/issues/17