Get base ID from string when opening detail pane
When opening a new detail pane location, get the id attribute of the
node from the string 'detail_chm' instead of from the id of the node
being cloned.

This change is made to facilitate the save-state feature. Prior to
starting development of the save-state feature, the node being cloned
had id 'detail_chm'. However with the current draft implementation
of the save state feature, the node being cloned can have and id such as
'detail_chm2', which leads to the resulting clone having id
'detail_chm23'.