Call updateSelection() after reconstruction
In order to ensure the primary detail pane is draw after reconstructing
the layout from a saved state, added a call to updateSelection(). At the
moment this call is wrapped in a setTimeout, and without that setTimeout
it doesn't appear to work (i.e. the primary detail pane has the labels
and outline, but the map itself is blank).

Additionally, things don't seem to work unless the primary detail pane
is drawn before the other detail panes. So that change is also in this
commit. (See changes to setPanesContent()).

This is a bad hack and needs to be fixed.