Call updateSelection on all DetailMaps in StateMan
In 4428d905 updateSelection() was called on the primary detail map as
a bad hack to ensure the map area was drawn after reconstructing from a
saved state. This commit applies that same hack to all the detail maps.
Without this, if there are several detail maps, some are blank (labels
and outline drawn, but the map area is blank).

This is a bad hack and needs to be fixed.