Bug fix: closing primary detail & creating another
This commit fixes a bug introduced in 67abee99: if the user closed
the primary detail map, then tried to create additional detail maps, the
code would crash.

The fix is to introduce another quasi-global variable analogous to the
'firstSwitch' variable removed in 67abee99. This new variable is
quasi-global in order to work with the reconstruction.