Only remove old dynamic labels if they exist
The original code logged a console error if the code tried to remove
a dynamic label that did not exist. I'm not sure if that was to help
with debugging the original code (before the save-state feature) or to
signify a more serious issue.

I changed the code to only remove the child div if it is indeed a child
element of the parent. If we find problems later, we can look here.