Change 'caption' to 'description' in ShaidyMapGen
The goal here is to obtain language consistency for both users and
developers. On the NGCHM viewer, the text in 'chm.info.caption' is
called a 'Description'. The plan is to also change the NGCHM-R package
to use 'chm.info.description' for this special property.

Backward compatibility is maintained for maps using 'caption'.