Bug fix to 2.20.3: mouseout event on summary map
On the current production release, 2.20.3, if there are no detail panes
open and a mouseout event occurs on the summary pane, there are errors
in the console. This commit addresses those errors by only calling
the updateSelection function if there are detail maps open. Hopefully
this 'fix' doesn't cause other problems.