Bug fix: multiple primary panes in DetailMaps
A big struggle in the save-state feature request has been the situation
where the user saves a panel layout without any detail maps open. This
was mostly addressed in 42628c4. However because of the block of code
added to DetailHeatMapDisplay when
	 NgChm.RecPanes.savedInitialDetailPane != undefined
there are sometimes two primary panes added to the NgChm.DMM.DetailMaps
array. This commit makes a bad hack to remove the extran primaryMap from
the array.