Bug fix: #161
Rationale for the two changes of this commit:

The RecreatePanes code used 'ngChmContainer1' as its base container,
resulting in the reconstructed state having TWO elements with id =
'ngChmContainer'.

In resizeNGCHM(), the variable topContainer is defined outside
the function. Since the div with id = 'ngChmContainer' is removed and
recreated as part of the panel reconstruction, the variable no longer
points to an actual DOM element after reconstruction. This resulted in
the oddities seen in issue #161.