Possible bug fix: #160
The if block changed in this commit is sometimes, but not always,
triggered during reconstruction of a saved state.

It's unclear to me from the code and commit history if the console
error and log removed in this commit is left over from debugging, or
if they were designed to alert to downstream problems.

Changed them to instead return, but will be on the look out for problems
this may cause.