Bug fix: click on pane icon not working
A new bug was introduced today: the pane menu was not showing up when
clicking the pane icon. This was because, similar to
0315eaf89c30b89ec18065ffae8c93c9167a2c9c, the DOM element with id =
ngChmContainer is removed and recreated during state reconstruction.
Therefore after reconstruction the NgChm.UTIL.containerElement variable
doesn't point to an actual DOM element after reconstruction. This commit
addresses that issue.