Bug fix: pane1/2 mapConfig info removed on load
The bug: If the map is loading from a saved state, and that saved state
has either a plugin or linkout in pane1 or pane2, the data needed to
reconstruct the pane is removed from mapConfig during a call to
NgChm.Pane.emptyPaneLocation().

The commit removes the call to emptyPaneLocation during the first calls
to switch to detail or summary in order to remove this bug.

Initial testing suggests this doesn't cause any problems afterward as
the user manipulates panes.