Get panel info from mapConfig before state reconst
Get the panel_configuration info from mapConfig before doing any of the
state reconstruction. Rationale: emptyPane removes information from
mapConfig for the panes it is called for. If we don't get this
information before starting reconstruction, the information for pane1
and pane2 will be destroyed before we can capture it.

PS: we want emptyPane to remove that information, so that as a user is
manipulating panes, information from closed panes is removed from
mapConfig.