Increase timeout for reconstructing detail maps
Without this increase, larger maps with multiple detail panes will use
higher-level tiles for the detail maps.

It seems like with a smaller timeout and large maps, the detail maps are
redrawn before all the needed tiles are loaded. Maybe the fix to this
hack of a function, triggerUpdateSelectionOnDetailMaps, is a Promise
that resolves when all the tiles needed for the multiple detail maps
have loaded.