Bug fix: Selections for plugins in RecPanes
If data on the NGCHM was selected via a plugin (e.g. by lassoing a
region on the 2D scatter plot), that data will NOT be saved in the
mapConfig.json file for the plugin,
<paneid>.data.axes[0].selectedLabels, when the user saves the .ngchm
file. This commit adds a bit of a hack to put in the current selections
in RecPanes so that the reconstructed pane plugins have the correct
selections.