Save only a single 'currentDl' to map config
At the moment in the code, each detail map has it's own data layer.
But we are going to change that so when the data layer toggle is changed
the data layer changes for all the detail maps. In preparation for that,
changing the saved data layer to the overall map data layer.