Bug fix: initializing plugin data in RecPanes
The initial implementation of sending saved data to plugins had some
bugs addressed here. The biggest issue was if a saved map was loaded,
and later the user changed one of the panels to a plugin, the function
RecPanes.initializePluginWithMapConfigData would be triggered, which is
not the desired behavior (this function is desiged only for initial
reconstruction.

The key fix here is to delete information from the mapConfig after it
has been used in reconstructing the saved panel layout. However other
minor improvements are checking for paneInfo existence and type in
RecPanes.initializePluginWithMapConfigData, and to change a variable
name for clarity.

Addresses #189