Add back hack to redraw detail maps in RecPanes
I had thought we could get rid of the terrible hack of calling
NgChm.SEL.updateSelection in a setTimeout (611661f4) by using a single
call to NgChm.SEL.updateSelections(true). But then there is issue #201.

This commit creates a similar another terrible hack that addresses the
issue...but ouch, it's not pretty.