Handle WebGL context loss and restore events
The browser can remove access to a WebGL context at basically any time.
When the context is removed, any WebGL objects for that context are
destroyed.

Until the context is restored, no WebGL functions for that context can be
used.
After the context is restored, no previous WebGL objects can be used.

To facilitate handling of the context loss and restore events and simplify
our use of WebGL, this patch creates and uses a common GLManager object
for each canvas.  GLManager is implemented in Drawing.js.