Add some basic error messages to the Widgetizer scanner
The Widgetizer will crash if any of the indexOf uses returns -1 because of
a badly formatted string in the input file.

Rather than expecting the user to debug the stack dump, we output some
very rudimentary error messages to the console.

TODO: Terminate the process more cleanly on error
TODO: Add similar error messages to the other apps without duplicating
      the code completely.