Workaround bug(s) redrawing restored context in Chrome/Safari on Macs
In Mac Chrome and Mac Safari, a canvas for which the WebGL context has
been lost and restored is not redrawn using WebGL operations.  The
operations on the canvas succeed, but the screen is not updated.

The workaround in this patch momentarily hides the canvas when the
context is restored.  This appears to be sufficient to revive the canvas.

Bug observed and workaround tested on Mac Chrome 96.0.4664.110 and Mac
Safari 14.1.2.  Mac Firefox 95.0.2 and Linux Chrome 95.0.4638.54 do not
seem affected by the issue.