Fix widget bugs introduced by commit 0592b5095845360f4162a670333b0f5181ec7891
That commit introduced a simpler way of determining the correct image
to use for the zoom mode buttons in the normal, hover, and selected states.

However, that broke the Widgetizer.  The Widgetizer requires images to
have full pathnames, which it converts to 'data:' URLs that are included
in the widget javascript code to make it standalone.

This patch adds a simple table lookup to convert from image names to
image paths, thus allowing the Widgetizer to do its conversions, which
also achieving the goals of the original patch.