Reduce widget size by reusing detail panel mode button images
The widgetizer replaces repeated image sources with repeated data sources,
unnecessarily increasing the size of the widget.

We move the function for initializing the mode buttons to the module
containing the mode button image changing functions and reuse the table
of image buttons, thus avoiding duplication of the data sources for the
buttons concerned.