Save selections to mapConfig during select drag
Save the selections to mapConfig.panel_configuration.selections during
a shift + drag event on the detail map (i.e. in
NgChm.DEV.handleSelectDrag).

I tried putting the call to NgChm.MMGR.saveSelectionsToMapConfig in
NgChm.DEV.clickEnd, but if the user scrolls off the edge of the map when
selecting, clickEnd doesn't fire, and therefore in such a situation the
selections wouldn't be saved.

If this becomes some sort of bottleneck, we can create a 'dragEnd' event
to handle activities such as this that should happen at the end of a
drag.

Addresses #204