Bug fix: copy plugin-name when closing pane
In d2f870eb we started using the data-plugin-name to store the plugin
name. However if the user closed a pane above the plugin's pane, that
attribute was not added to the resulting pane containing the plugin,
which lead to #263.

The commit addresses the bug by copying the data-plugin-name attribute
to the new pane.