Bug fix: draw labels during zoom in after zoom out
Prior to 6581d45, NgChm.DET.resizeOnNextDraw was set to true in every
pass through NgChm.DET.setDrawDetailTimeout. After this was changed to
instead set mapItem.resizeOnNextDraw = true based on the value of the
noResize argument, issue #261 appeared.

This commit addresses #261 by setting noResize to false in the function
that handles the zoom in.