Bug fix: top items not being displayed
This is a bug fix for issue #277: top items for rows and columns were
not being shown. This bug was introduced in 59d2d061 by removing the
function call to NgChm.SUM.summaryInit() in NgChm.UPM.prefsSuccess().
(NgChm.SUM.summaryInit() is the function that adds the top items to
NgChm.SUM.rowTopItems and NgChm.SUM.colTopItems.)

That function has since changed names:
  NgChm.SUM.summaryInit() -> NgChm.SUM.initSummaryData()

I notice that commit 59d2d061 was made to address issues when the saved
state has zero detail and/or zero summary panes. I verified that the
change in this current commit works when there are zero detail and/or
zero summary panes (i.e. tested with just a plugin pane in the saved
state).