Bug fix: prefs pane jumps to mouse on drag
This addresses a long-standing issue where at the start of a drag of
the Heat Map Display Properties dialog, the dialog jumps so that its top
left corner is at the mouse location.

To the best of my knowledge, we don't have a GitHub issue for this, but
we've all noticed it.