Center message box horizontally
Prior to this commit, if the user dragged the message box around, closed
it, and opened another message box, the newly-created message box would
have the same horizontal positioning as the closed message box, but would
have vertical placement based off the 'mdaServiceHeader' element.

So here added a function to display the message box, and set the
placement.