Bug fix: scrolling user prefs when dialog is small
Adding this quick and dirty implementation to address the following
issue: if the user preferences dialog is not very tall, the user cannot
scroll the inner div to see the very bottom items.

This is a rough hack, and we need to take the time to address layout
issues and implementation in this dialog.