Keep additional HTML tags when making widget
Issue #249 occurs because the code is looking for an element with
id="viewport". The chm.html file has this in the <head>, but the
widgetizer doesn't keep those lines. It used to--for example f6b6c7b
does NOT have this bug.

The current commit makes changes to keep these lines and therefore
address #249.