Bug fix: viewport in GUI builder
This is a kludge to address issue #249. Note that cbd2415 does not
actually address the issue for the GUI builder, because at the time of
writing, the GUI builder does not use NGCHM_Embed.js.

Background info:
Issue #249 occurs because the code is looking for an element with
id="viewport". The chm.html file has this in the <head>, but the
widgetizer doesn't keep those lines. It used to--for example f6b6c7b
does NOT have this bug.