Make plugin names on help table links
Before only the logo was a link to the website. But now we are removing
logos for links that don't work. In order to let users know there is
(or at least was, or at least sometimes) is a website for the plugin,
make the name also a link to the website.