Move DET.minLabelSize to UTIL.minLabelSize
minLabelSize is not used in the DET module.
Since it isn't specific to any other module either and does not depend on
any other module, I think it belongs in the UTIL module.

At some point this function was duplicated (one in UTIL, one in SUM). The
one in SUM is not used and it has nothing to do with SUM, so I removed
that one.