Call CUST.addCustomJS only after NGCHM initialized
Currently, MatrixManager calls CUST.addCustomJS after the mapConfig
object is received but possibly before the mapData object is
received and the heat map is fully initialized.  CUST.addCustomJS
checks to see if the map is fully initialized and, if not, waits for
an arbitrary 100ms in the hope that by then the map will have been fully
initialized.  For big maps over slow networks, it is very possible
that this hope is dashed and CUST will attempt to add custom.js before
the map is initialized (and this will fail).

This patch replaces this process by having the UI Manager initiate
loading of custom.js after it has received the MMGR.Event_INITIALIZED
message.

Consequently, there is no possibility that the map is not initialized
before addCustomJS is called and no possibility addCustomJS will fail
because the map is not initialized.  We can also remove the entire event
callback "loop" in addCustomJS that was waiting for the map to be initialized.

And we can remove the dependency of MMGR on CUST, since CUST.addCustomJS
is now called from UI-Manager.js in response to the Event_INITIALIZED
event.