Add DEV.keyNavigate after map is initialized.
Currently, the DEV.keyNavigate event listener is activated when the
mapConfig object is received but possibly before the mapData object
is received and the heat map is fully initialized.  This can cause
errors if keys are pressed while the map is loading.

This patch moves adding the DEV.keyNavigate event listener until after
the map is fully initialized by using the Event_INITIALIZED listener in
UI-Manager.js.

This also allows us to remove the dependence of MMGR on the DEV module.