Move UI fns for saving maps from MMGR to UIMGR.
Move UI-related functions associated with saving maps from MMGR to UIMGR.
In order to save the current map state, many high-level modules must be
queried to obtain the required details.  Including or calling those
functions in MMGR creates circular dependencies between MMGR and the
modules concerned.  The UIMGR can, by definition, query all such
modules.  So this patch moves the map save capability to the UI manager.

For autosave on compatibility update, this required moving the related
code to the EVENT_Initialized callback in the UIMGR and providing a way
for it to determine if a compatibility update was needed.

This patch removes dependence on LNK, PANE, SUM, DMM, DET, and SRCH
from MMGR, and removes dependence on MMGR from CompatibilityManager
(CM).

Functions now in UI-Manager between the comments
    /***
    *  Functions related to saving Ng-Chms.
    */
and
    // End of map save functions.
have basically been moved from MMGR to UI-Manager.