Move redrawCanvases from UIMGR to SUM. Remove deps on UIMGR from DEV, PDF, UPM.
redrawCanvases is mostly about redrawing the canvasses in SUM, but it
also adds an event to redraw a canvas in DET.  It's possible that SUM will
need to depend on DET anyway, so moving it to SUM will do for now.

I also considered splitting it into two (one each for DET and SUM) but I
decided that opens the door to confusion about whether to call just one
or both.

I'm also not convinced this is not the right way to address the issue if
it even still exists.