Count and trace modules imported before they are defined
If, and it is a big if, the modules are imported in the order
that minimizes the number of such out-of-order imports, this
is equivalent to determining the minimum number of edges that
must be removed to convert the module graph into an acyclic graph.

In general, this is the minimum feedback arc set problem in
graph theory and it is NP-hard.  However, all the complexity
involved is in determining the optimal order of imports (there
are also other approaches).

I think we can assume that the import order is already at least
a very good approximation to the best order and good enough for
our needs.  And a lot simpler to implement.

The code is included in a region of NgChm.js that will be excluded
from the production code by the compiler.