BUGFIX: Detect subDendroMode again
This is a start towards fixing a bug that was made in commit c5f05b5bcca9a
on 2019-07-11.  The code was committed by Mark Stucky but it was my code.
(At this time, Mark was collecting all the patches etc. and then
committing them to github in batches.)

The commit concerned was when I totally rewrote the dendrogram code.

The previous code was setting the 'DDR.subDendroMode' global variable
to row or column if the user clicked on a summary view dendrogram branch.
The replacement code never set the variable.

The variable has been tested in DetailHeatMapDisplay ever since, failing
every time.  The test is supposed to select row or column scaling
on one axis only when the user zooms all the way out in subDendro mode.
It wasn't. (Both axes were being scaled, as when zooming all the way
out in normal mode.)

This patch does reenable that feature.  However, it does not properly
reset the scaling when the user subsequently zooms in again.