Split PluginInstanceManager.js (PIM) out of Linkouts.js
Except for a couple of exceptions detailed below, the contents of the
new PluginInstanceManager.js used to be inside an IIFE within
Linkouts.js.  Splitting them out into a separate modules allows us
to break the circular dependence between the Linkouts and Search
modules.

The first exception is that the function LNK.postSelectionToLinkouts
was not within the IIFE but was moved to the new module anyway.
Indeed, it is the crucial function that enables the circular dependence
to be removed.  It does not depend on any other logic within Linkouts.js.
The function was also renamed to PIM.postSelectionToPlugins to be
consistent with the other functions in the module.

The second exception is that the function LNK.addLinkoutPlugin was not
moved to PIM even though it was originally within the IIFE.  This
function is closely tied to the linkouts module and is just a client of
the functions that have been moved to PIM.

A third function worth commenting on is PIM.setPanePluginOptions.  This
function modifies the plugin instances data structure, so it should be in
PIM.  However, its last step was to call LNK.initializePanePlugin to
get the data to send to the plugin.  This function clearly belongs in
Linkouts.js.  So, rather than create a (circular) dependence on LNK,
LNK.initializePanePlugin is passed as a function parameter to
PIM.setPanePluginOptions.