Move NGCHM representational constants into a new module HeatMapRep.js
The ColorMapManager needs some of these to draw color maps, but
the MatrixManager has to create ColorMapManagers, so that's
a fundamental circular dependency.  We break it my moving
the represention constants into their own module.

In the future, I imagine we can move other basic functions for interpreting
the contents of a .ngchm file into this module as well.