Rename SelectionManager.js to DetailHeatMapViews.js (detail map reorg part 1)
SelectionManager has always been a very confusing name to me. When I
think "Selections" I think about the labels that get highlighted when
when you click on them etc.  Indeed, the first several times I tried
writing that last sentence, I kept saying "selected labels" or something
to that effect.

Anyway, every function in the module except a few is actually operating
on a mapItem, i.e. a detail map view.  So, I've decided to rename this
module (SelectionManager.js) to DetailHeatMapViews.js.

I also think it will be a good place to move additional DetailHeatMap
material on which both the DetailHeatMap and SummaryHeatMap modules
depend.

This commit just renames the file to make it easy to see in the git
history.  The next commits make the code updates required.