Remove unused variables DVW.currentRow and DVW.currentCol
These variables were used to indicate the top-left corner of the
detail pane before the change that gave us multiple detail views.
Now they're set but not used.

They were also used to set the position of the detail view on
startup from URL parameters.  That functionality apparently has
not worked since that time.  And no one has missed it.

It is not clear what setting those URL parameters should mean
now that we have saved state.  (Probably: override saved state
of primary detail window if it exists.)

Since it hasn't worked/been used for so long, I've just deleted
the last vestiges of that functionality.  Easy enough to reimplement
if we decide it's needed (and what it means).