Pull out SearchState.js from SearchManager.js
This commit separates access to the search state from
modifications to the search state.  Having them in the
same module forces circular dependencies: e.g. SummaryHeatMap
and DetailHeatMap want to highlight the search results on
their displays, but changing the search requires the
summary and detail heat maps to be redrawn.

SearchState.js is for read-only access to the search results.
All updates to the SearchState should go through the SearchManager,
which will ensure that all consequent changes are performed.
This latter requirement is not enforced.