Define embedding functions in UI-Manager but API namespace
The embedding functions belong in the API namespace (so we don't
expose the internal module structure of NGCHM viewer to external apps).
But these functions require access to the NGCHM loading functions that
are currently defined in UI-Manager.   Perhaps, it is worthwhile
pulling all those functions out of UI-Manager into their own module.

But short-term solution for now is to add these embedding functions into
UI-Manager