Fix git commit hash embedding
Fix injection of git hash into ngchmEmbed-min.js and ngchmWidget-min.js.

Previously git could not determine the hash because the .git directory
was not being copied into the build environment.  This failure was
causing ant to output a 128 error code and an error message to be
injected into the above modules instead of the lastest git hash.