Fix issue #348 - unable to download viewer
Original issue was caused by a simple scoping problem.  However, fixing
it involved accessing fileSrc, which is a function parameter that was
assigned to global variable in the MMGR module.  If we ever support
multiple heat maps, this will need to be specific to individual heat
maps.

So as part of this commit, I also made the few changes required to
make it a heat map specific property.