Fix issue #356 - unable to deselect labels
Cause was clearAxisSearchItems was not reexported from SearchManager.

But a bigger issue was that SummaryDendrogram was calling search result
modification functions directly instead of via SearchManager.

To correct that I had to reorganize a bit and pass the desired search
result modifications in via ddrCallbacks, which had to be defined using
callbacks passed into DetailHeatMapDisplay from UI-Manager.