Fix issue #369 - server.app cannot load maps from the server
The closure compiler used to minify the javascript used by the server.app
mangles/removes the names of functions, including those used sent to the
worker app used to download assets from the server.  The compiler does
not (cannot) also update these names used in string constants.

Replace the names of these functions in string constants with the names
assigned to the functions by the closure compiler.