Update ribbon mode dendrogram selection when changing detail maps
Updates the behavior of the ribbon mode dendrogram selection as we
discussed.

Tracks whether each detail view's ribbon mode focus area is a dendrogram
selection. Initially false. Set to true if the user selects a dendrogram
branch on the summary map.  Cleared if the user uses keyboard navigation
to change the focus area (and displays a popup allowing the user to reset
the focus region if desired).

Also used to update dendrogram ribbon mode highlight when switching
between detail maps or when restoring saved maps.

Fixes issue #372.