Add preliminary support for varying size of detail map covariate bars
Adds two new mapItem fields: rowClassScale and colClassScale that control
the size of the covariate bars relative to the size of the detail map.
These are constants for now. A future commit will provide a way to
adjust these scales so that the absolute size of the covariate bars is
less variable with respect to mode changes and panel size changes.

The previous heatMap methods for accessing the covariate bar heights have
been replaced a function that returns an array of scaled covariate bars.
The array has additional methods for computing summaries of the scaled
covariate bars.

For convenience computing scaled covariate bars for detail maps, the
getScaledVisibleCovariates method has been added to mapItems.  It
automatically uses the mapItem scaling for the covariate's axis.

In addition, I have replaced many pairs of functions that came in row
and column versions with a single function that takes the axis as a
parameter.