Consolidate changes to data view widths/heights to a single function.
This function also adjusts the scaling of the mapItem's covariate
bars so that they stay roughly the same size with respect to the
heat map when the data view size changes.

To see the difference, open a large map with many covariate bars and
switch between normal mode and ribbon mode.