Fix issue #101. Implement special listener for AccessWindow.onready().
Adds a single eventListener to the heatMap to check for AccessWindows (or
TileWindows) that are waiting for all tiles to be available.  Once a
waiting onready() listener has been called, it will be removed from the
list of AccessWindows to check and so it cannot be called multiple times.

Problems with the previous implementation included:
- Multiple calls to target function when tiles requested & loaded multiple
  times.
- Multiple calls to target function once the AccessWindow is complete but
  tiles outside the AccessWindow are loaded.
- Multiple calls to target function if we ever expunge tiles and then
  reload them.