Fix error where detail maps would sometimes not fetch detail data
I saw this when switching between layers on a large map (Deprez3k) with
a detail panel open showing a horizontal ribbon view.  The view would
update to an approximate version (rows of identical data) but not update
to the detail data.  I believe this was because the detail tiles concerned
were never request.

This commit makes creates an AccessWindow for all detail maps whenever
they redraw.  The access window will check if all the tiles used by the
window are available and request any that aren't.