Remove excessive yet insufficient data tile fetching when creating PDF
Previously, the browser would request and wait for all tiles in the data
level of the primary detail view.  This is excessive since the only tiles
needed are those used by that view and other views at the same level. If
you are looking at a large map and just have one small detail view, loading
every detail tile in the NG-CHM is excessive, causes a long delay, and
loads a lot of unneeded tile data into the browser.

It is also insufficient if you have detail views at different levels open;
for example, a normal detail view and a ribbon view. Data will be fetched
for whatever level is in the primary view but not for the other view.
Fortunately, it is more than likely the browser already has all the
necessary tiles available.  When you selected/moved each detail view,
the browser requested all the tiles for that view.  Unless your network/
system is very slow and you're very fast to start generating the PDF, the
browser will have loaded all the necessary tiles.

This commit replaces the single large data tile request with a separate
request for just those tiles needed for each detail view.  The main effect
is to remove the unnecessary request for all the detail tiles.  It is very
hard to observe the requests for each individual detail window.  I was
able to by
- opening a large multi-layer NG-CHM (e.q. Deprez3k),
- creating at least three large detail, h ribbon, and v ribbon views at different
parts of the map,
- quickly switching to the second data layer, and
- *immediately* generating a PDF of just the detail views.
I was never quite fast enough until I also used Chrome's performance tools
to slow down the browser to low-end mobile.

The change to the new asynchronous architecture required some code cleanup
and I also did a little just for the sake of it, but there is still a lot
left to do.